digital_cookbook/requirements.txt
flask
pymysql
cryptography
psycopg2-binary
